refactor(vector): move metric_for_opoid to VectorMetric::from_opoid
Co-locate the operator-OID -> metric resolution with the other VectorMetric conversions (opclass_name, operator, from_opfamily_name) instead of a free function in orderby.rs. Behavior unchanged (still guards on to_regtype so pgvector need not be installed).
