+++ pg_search/src/vector/metric.rs
@@ -30,7 +30,7 @@
 //! tantivy schema's `VectorOptions::metric`.
 
 use crate::postgres::catalog::lookup_opfamily_name;
-use pgrx::pg_sys;
+use pgrx::{direct_function_call, pg_sys, IntoDatum};
 use serde::{Deserialize, Serialize};
 use tantivy::vector::Metric as TantivyMetric;
 
@@ -69,6 +69,42 @@ impl VectorMetric {
             VectorMetric::Cosine => "<=>",
             VectorMetric::InnerProduct => "<#>",
         }
+    }
+
+    /// Map a pgvector distance operator OID (`<->`, `<=>`, `<#>`) to its
+    /// implied metric. Returns `None` when `opoid` is not one of them —
+    /// which includes the case where pgvector is not installed at all.
+    ///
+    /// The operator OIDs are resolved once and cached. Resolving them via
+    /// `regoperatorin` raises `type "vector" does not exist` when pgvector is
+    /// absent (a Postgres ereport, not catchable by `unwrap_or`), and this runs
+    /// on every custom-scan plan — so gate the lookups on the `vector` type
+    /// existing (`to_regtype` returns NULL rather than erroring); absent it, all
+    /// OIDs stay INVALID and no metric ever matches.
+    pub(crate) fn from_opoid(opoid: pg_sys::Oid) -> Option<Self> {
+        use std::sync::OnceLock;
+        static OP_METRICS: OnceLock<[(pg_sys::Oid, VectorMetric); 3]> = OnceLock::new();
+        let cached = OP_METRICS.get_or_init(|| unsafe {
+            let vector_type_exists =
+                direct_function_call::<pg_sys::Oid>(pg_sys::to_regtype, &["vector".into_datum()])
+                    .is_some();
+            let lookup = |sig: &std::ffi::CStr| -> pg_sys::Oid {
+                if !vector_type_exists {
+                    return pg_sys::Oid::INVALID;
+                }
+                direct_function_call::<pg_sys::Oid>(pg_sys::regoperatorin, &[sig.into_datum()])
+                    .unwrap_or(pg_sys::Oid::INVALID)
+            };
+            [
+                (lookup(c"<->(vector,vector)"), VectorMetric::L2),
+                (lookup(c"<=>(vector,vector)"), VectorMetric::Cosine),
+                (lookup(c"<#>(vector,vector)"), VectorMetric::InnerProduct),
+            ]
+        });
+        cached
+            .iter()
+            .find(|(oid, _)| *oid != pg_sys::Oid::INVALID && *oid == opoid)
+            .map(|(_, metric)| *metric)
     }
 
     /// Map a pg_opfamily name (e.g. `vector_cosine_ops`) to its implied
